- Add Dto LogItem
- Added property LogItem
- Create new Entity Logitem
- Added Logitem Property
- Added new filter for global exception handling
- Added Filter in controller
- Add StatusCode,TraceId and Error
- Create ValidationError  Class
- Create Logitem  entity
- Create Error Handling Filter
- Create Validation Failed Result
- Update Program.cs
- Merge branch 'develop' into 2866-R&D_for_Global_Exception_Handling_in_Backend
- changes in error response
- updated mapping for Campus in Department
- Create GetDepartmentByCampus API
- Added CampusId relation in Department table
- updated Department Service for Creating Department with CampusId
- updated payrollPayment process for creating payment with campusId
- Updated payrollTransacitonService for creating Payroll by campus
- Updated mapping for absent Days in PayrollTransaction
- updated Dtos
- Added CampusId to DeptFilter for payroll
- Added Filter for payroll report with campus
- Added Response method for errors
- added DropDown method for DepartmentByCampus
- Updated specs
- added specs for fetching only campus which are active
- Removed CampusId relation with Employee
- updated CampusName and CampusId mapping
- Remove CampusId from CreateDto
- updated Specs
- Merge remote-tracking branch 'origin/develop' into 2866-R&D_for_Global_Exception_Handling_in_Backend
- Updated Service and Dto
- Updated Dto
- remove CampusId comments
- Added Migration for Campus in department
- Merge remote-tracking branch 'origin/develop' into 3109-Add_campus_in_Department_API_and_Payroll_Process
- Add Type casting
- Fix issues
- merging with 2866 Exception Handlings
- added Explicit TypeCasting
- Added Nullables to All required attributes in CreateDtos
- Added StatusCode to Response
- Added LogItem Migration for Exception handling

Related work items: #3109

+++ Application/Services/BankReconService.cs
@@ -43,20 +43,20 @@ namespace Application.Services
 
 
             // FOR BANK STATEMENT
-            var bankStmtLine = await _unitOfWork.BankStmtLines.GetById(entity.BankStmtId, new BankStmtLinesSpecs());
+            var bankStmtLine = await _unitOfWork.BankStmtLines.GetById((int)entity.BankStmtId, new BankStmtLinesSpecs());
             decimal stmtAmount = bankStmtLine != null ? (bankStmtLine.Credit - bankStmtLine.Debit) < 0 ? (bankStmtLine.Credit - bankStmtLine.Debit) * -1
                 : bankStmtLine.Credit - bankStmtLine.Debit : 0;
 
-            var reconciledStmtAmount = _unitOfWork.BankReconciliation.Find(new BankReconSpecs(entity.BankStmtId, false)).Sum(i => i.Amount);
+            var reconciledStmtAmount = _unitOfWork.BankReconciliation.Find(new BankReconSpecs((int)entity.BankStmtId, false)).Sum(i => i.Amount);
             var unreconciledStmtAmount = stmtAmount - reconciledStmtAmount;
 
             if (entity.Amount > unreconciledStmtAmount)
                 return new Response<int>("Amount is greater than Unreconciled Statment Amount");
 
             // FOR PAYMENT
-            var payment = await _unitOfWork.Payment.GetById(entity.PaymentId, new PaymentSpecs(true));
+            var payment = await _unitOfWork.Payment.GetById((int)entity.PaymentId, new PaymentSpecs(true));
             decimal paymentTotalAmount = payment.GrossPayment - payment.IncomeTax - payment.SalesTax - payment.SRBTax;
-            var reconciledPaymentAmount = _unitOfWork.BankReconciliation.Find(new BankReconSpecs(entity.PaymentId, true)).Sum(i => i.Amount);
+            var reconciledPaymentAmount = _unitOfWork.BankReconciliation.Find(new BankReconSpecs((int)entity.PaymentId, true)).Sum(i => i.Amount);
             decimal unreconciledPaymentAmount = paymentTotalAmount - reconciledPaymentAmount;
 
             if (entity.Amount > unreconciledPaymentAmount)
@@ -70,8 +70,8 @@ namespace Application.Services
                 await _unitOfWork.SaveAsync();
 
                 //FOR BANK STATEMENT STATUS
-                decimal reconciledTotalStmtAmount = _unitOfWork.BankReconciliation.Find(new BankReconSpecs(entity.BankStmtId, false)).Sum(i => i.Amount);
-                var bankStmtLineforUpdate = await _unitOfWork.BankStmtLines.GetById(entity.BankStmtId);
+                decimal reconciledTotalStmtAmount = _unitOfWork.BankReconciliation.Find(new BankReconSpecs((int)entity.BankStmtId, false)).Sum(i => i.Amount);
+                var bankStmtLineforUpdate = await _unitOfWork.BankStmtLines.GetById((int)entity.BankStmtId);
 
                 if (bankStmtLineforUpdate != null)
                 {
@@ -86,7 +86,7 @@ namespace Application.Services
                 }
 
                 //FOR PAYMENT STATUS
-                decimal reconciledTotalPayment = _unitOfWork.BankReconciliation.Find(new BankReconSpecs(entity.PaymentId, true)).Sum(i => i.Amount);
+                decimal reconciledTotalPayment = _unitOfWork.BankReconciliation.Find(new BankReconSpecs((int)entity.PaymentId, true)).Sum(i => i.Amount);
 
                 if (paymentTotalAmount == reconciledTotalPayment)
                 {
